Dashboard works correctly. Fetches and displays article units, vehicles and orders correctly for the correct corresponding warehouse.

+++ backend/src/main/java/mk/ukim/finki/db/distributorapp/model/dto/WarehouseInventoryDto.java
@@ -8,5 +8,5 @@ import lombok.Data;
 public class WarehouseInventoryDto {
     String articleName;
     String manufacturerName;
-    Integer totalUnits;
+    Long totalUnits;
 }
